通过扩展类来扩展AbpSession（推荐此种做法）

+++ LearningMpaAbp.Application/Tasks/Dtos/CreateTaskInput.cs
@@ -4,7 +4,7 @@ namespace LearningMpaAbp.Tasks.Dtos
 {
     public class CreateTaskInput
     {
-        public int? AssignedPersonId { get; set; }
+        public long? AssignedPersonId { get; set; }
 
         [Required]
         public string Description { get; set; }
